Add "Full Screen" mode to Sample app
